arreglando error en destroy.sh #69

+++ destroy.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
 
-kill -9 `ps -ef|grep -v grep |grep dev_appserver | awk '{print $2}'`
-echo 'El servicio esta funcionando'
+line=$(ps -ef|grep -v grep |grep dev_appserver | awk '{print $2}')
+if [[ $line -gt 0 ]];
+    then
+    kill -9 `ps -ef|grep -v grep |grep dev_appserver | awk '{print $2}'`
+else
+	echo 'El servicio no está funcionando. Omitiendo...'
+fi
